fix(presentations): tira remark-prism do bundle client
/presentations quebrava em produção com SharedArrayBuffer is not defined
porque remark-prism/jsdom ia pro client junto com o componente. getStaticProps
agora vive em lib/presentations-static-props.ts — bundle cai de 1.42MB pra 328kB.
